+++ R/cell_types.R
@@ -29,7 +29,7 @@
 #'
 #' @export
 update_metacell_types <- function(project, dataset, metacell_types_file) {
-    verify_project_dir(project)
+    init_project_dir(project)
     verify_app_cache(project)
 
     prev_metacell_types <- load_shiny_data("metacell_types", dataset, project_cache_dir(project))
@@ -85,7 +85,7 @@ update_metacell_types <- function(project, dataset, metacell_types_file) {
 #'
 #' @export
 update_cell_type_colors <- function(project, dataset, cell_type_colors_file) {
-    verify_project_dir(project)
+    init_project_dir(project)
     verify_app_cache(project)
 
     cell_type_colors <- parse_cell_type_colors(cell_type_colors_file)