+++ inst/app.R
@@ -1,7 +1,14 @@
 # Launch the ShinyApp
 
 options("golem.app.prod" = TRUE)
-if (dir.exists("code")){
+if (dir.exists("code")) {
     pkgload::load_all("code", export_all = FALSE)
 }
-MCView::run_app(project = "project")
+
+if (dir.exists("project")) { # here for backward compatibility
+    MCView::run_app("project")
+} else {
+    MCView::run_app(getwd())
+}
+
+